+++ launcher/Session.vb
@@ -124,6 +124,7 @@ Public Class Session
         start.Environment("DIRT2VR_WORLD_SCALE") = "1"
         start.Environment("DIRT2VR_HEADSET_SCALE") = (settings.HeadsetScale / 100.0).ToString(Globalization.CultureInfo.InvariantCulture)
         start.Environment("DIRT2VR_FOV_SCALE") = (settings.FieldOfView / 100.0).ToString(Globalization.CultureInfo.InvariantCulture)
+        start.Environment("DIRT2VR_HUD_FOLLOW") = If(settings.HudFollowView, "1", "0")
         ConfigureLogging(start, logFolder)
         start.Environment("DIRT2VR_INPUT_CHANNEL") = channel
         start.Environment("DIRT2VR_KEYS") = $"{settings.ToggleKey}:{settings.ToggleModifiers},{settings.RecenterKey}:{settings.RecenterModifiers}"